Backport fix coq/coq#14116 | PR 12164 | Stop relying on side-effects for recursive scheme declaration.
Instead, we register functions dynamically declaring the dependencies of the
scheme to be generated.

We had to fix the test-suite because an internal scheme name changed.
We could also tweak the internal flag of scheme dependencies, but in this
particular case it looks more like a bug from the previous implementation.