Backport fix coq/coq#14116 | PR 14391 | Mark all scheme dependencies as internal.
The only effect this flag has now is to prefix the name of the generated
lemma with "internal_" when the scheme function is called outside of a
tatic.

I had to fix the Rewriting Schemes flags because it was not explicitly
declaring the involutive_sym scheme, leading to an internal declaration.