Catch exceptions thrown by Ind_tables.find_scheme
In rare instances, this function throws an exception which is not wrapped into
the tactic monad.

For unknown reasons, all known instances of this issue were fixed with the
introduction of https://github.com/coq/coq/pull/13563 in Coq 8.14. Hence, we do
not upstream this patch until the bug becomes visible again.